requirements.txt
asyncio-nats-client==0.11.4
